Update, tenji mini-test.
