+++ mini_test/mini_test_tenji.rb
@@ -2,20 +2,20 @@
 
 require 'minitest/autorun'
 require 'minitest/reporters'
-require './lib/tenji_minitest'
-
-lib = File.expand_path('lib', __dir__)
-$LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-
-Encoding.default_internal = 'UTF-8'
-Encoding.default_external = 'UTF-8'
+require './lib/tenji'
 
 class TenjiMakerTest < Minitest::Test
+  def self.encoding_style
+    Encoding.default_internal = 'UTF-8'
+    Encoding.default_external = 'UTF-8'
+  end
+
   def setup
-    @tenji_maker = TenjiMaker.new
+    @tenji_maker = TenjiMakerQiita.new
   end
 
   def test_a_hi_ru
+    encoding_style
     tenji = @tenji_maker.to_tenji('A HI RU')
     assert_equal <<~TENJI.chomp, tenji
       o- o- oo
@@ -25,6 +25,7 @@ class TenjiMakerTest < Minitest::Test
   end
 
   def test_ki_ri_n
+    encoding_style
     tenji = @tenji_maker.to_tenji('KI RI N')
     assert_equal <<~TENJI.chomp, tenji
       o- o- --
@@ -34,6 +35,7 @@ class TenjiMakerTest < Minitest::Test
   end
 
   def test_si_ma_u_ma
+    encoding_style
     tenji = @tenji_maker.to_tenji('SI MA U MA')
     assert_equal <<~TENJI.chomp, tenji
       o- o- oo o-
@@ -43,6 +45,7 @@ class TenjiMakerTest < Minitest::Test
   end
 
   def test_ni_wa_to_ri
+    encoding_style
     tenji = @tenji_maker.to_tenji('NI WA TO RI')
     assert_equal <<~TENJI.chomp, tenji
       o- -- -o o-
@@ -52,6 +55,7 @@ class TenjiMakerTest < Minitest::Test
   end
 
   def test_hi_yo_ko
+    encoding_style
     tenji = @tenji_maker.to_tenji('HI YO KO')
     assert_equal <<~TENJI.chomp, tenji
       o- -o -o
@@ -61,6 +65,7 @@ class TenjiMakerTest < Minitest::Test
   end
 
   def test_ki_tu_ne
+    encoding_style
     tenji = @tenji_maker.to_tenji('KI TU NE')
     assert_equal <<~TENJI.chomp, tenji
       o- oo oo
@@ -71,6 +76,7 @@ class TenjiMakerTest < Minitest::Test
 
   # single word mini-test.
   def test_A
+    encoding_style
     tenji = @tenji_maker.to_tenji('A')
     assert_equal <<~TENJI.chomp, tenji
       o-
